Fixed floating point issue

+++ src/MLTreeMaker.h
@@ -233,12 +233,12 @@ class MLTreeMaker: public ::AthHistogramAlgorithm {
     float m_TileBar1[4][4];
     float m_TileBar2[2][4];
 
-    float m_duplicate_EMB1;
-    float m_duplicate_EMB2;
-    float m_duplicate_EMB3;
-    float m_duplicate_TileBar0;
-    float m_duplicate_TileBar1;
-    float m_duplicate_TileBar2;
+    int m_duplicate_EMB1;
+    int m_duplicate_EMB2;
+    int m_duplicate_EMB3;
+    int m_duplicate_TileBar0;
+    int m_duplicate_TileBar1;
+    int m_duplicate_TileBar2;
 
 };
